added uli

+++ DSP-SRC/Source/inc/gui_interface_communications.h
@@ -1,8 +1,8 @@
 #include "main.h"
 
 typedef struct {
- uint8_t offset;
- uint8_t data[500];
+ uint32_t offset;
+ uint64_t data[500];
 } data_struct;
 
 void USART_puts(volatile char *s);
